fix: revert applyDefaultHierarchyTemplate in prism-native
The hierarchy template addition in e39a6f4 broke androidNativeWindow
cinterop visibility. The template promotes androidNativeMain into a
real shared source set, and the per-target cinterop registered on
androidNativeArm64/X64 leaf compilations is not visible from the
shared source set under kotlin.mpp.enableCInteropCommonization=true
when only one leaf cinterop has run.

A prior leaf-extraction attempt (4e4888b) also failed — the cinterop
was not visible from the leaf source sets either with the template
applied. Reverting to the manual nativeMain wiring restores the
exact configuration that worked in PR #49.

- prism-native/build.gradle.kts: remove applyDefaultHierarchyTemplate(),
  restore val nativeMain by creating + per-target defaultSourceSet
  dependsOn wiring
- prism-native/.../AndroidBridge.kt: restore the cinterop import and
  jniAttachSurface function (was removed in 4e4888b)
- prism-native/src/androidNativeArm64Main, X64Main: removed (no longer
  needed once shared androidNativeMain can see the cinterop again)
